+++ ARNurvousBreakdown/GameScreen/GameViewController.xib
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="15705" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15706"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="GameViewController" customModule="ARNurvousBreakdown" customModuleProvider="target">
+            <connections>
+                <outlet property="arView" destination="fwu-14-WM6" id="bp1-yY-elr"/>
+                <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
+            <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <arView multipleTouchEnabled="YES" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="fwu-14-WM6">
+                    <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                    <gestureRecognizers/>
+                    <connections>
+                        <outletCollection property="gestureRecognizers" destination="XKI-Pr-4XN" appends="YES" id="IQh-AC-l0Z"/>
+                    </connections>
+                </arView>
+            </subviews>
+            <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <constraints>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="fwu-14-WM6" secondAttribute="trailing" id="dMr-tO-Ccr"/>
+                <constraint firstAttribute="bottom" secondItem="fwu-14-WM6" secondAttribute="bottom" id="dQY-ga-izT"/>
+                <constraint firstItem="fwu-14-WM6" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="k5i-8q-Idc"/>
+                <constraint firstItem="fwu-14-WM6" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" id="yKV-KI-fPk"/>
+            </constraints>
+            <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
+            <point key="canvasLocation" x="139" y="100"/>
+        </view>
+        <tapGestureRecognizer id="XKI-Pr-4XN">
+            <connections>
+                <action selector="onTap:" destination="-1" id="50K-Lm-nbo"/>
+            </connections>
+        </tapGestureRecognizer>
+    </objects>
+</document>
